requirements.txt
biopython==1.79
numpy==1.20.3
pandas==1.4.1